frontend/backend/requirements.txt
fastapi==0.115.0
uvicorn==0.27.1
python-dotenv==1.0.0
httpx==0.24.1
pydantic==2.4.2